# docs/swagger.yaml
definitions:
...
  collectionmodels.Price:
    properties:
      basePrice:
        type: number
      price:
        type: number
      specialPrice:
        type: number
      vat:
        type: number
    type: object
  collectionmodels.ProductList:
    properties:
      productId:
        type: string
      quantityUse:
        type: integer
    type: object
  collectionmodels.ShopItem:
    properties:
      lazada:
        $ref: '#/definitions/collectionmodels.StandardShopItem'
      shopee:
        $ref: '#/definitions/collectionmodels.StandardShopItem'
    type: object
  collectionmodels.StandardShopItem:
    properties:
      itemId:
        type: string
      sellerSku:
        type: string
    type: object
  collectionmodels.Stock:
    properties:
      available:
        type: integer
      quantity:
        type: integer
      remaining:
        type: integer
    type: object
  models.AddStockI:
    properties:
      barcodeId:
        type: string
      productId:
        type: string
      quantity:
        type: integer
      updateType:
        type: string
    required:
    - barcodeId
    - quantity
    - updateType
    type: object
...
  models.CreateProductI:
    properties:
      barcodeId:
        type: string
      basePrice:
        type: number
      groupId:
        type: string
      isBaseProduct:
        type: boolean
      productName:
        type: string
      productNameTh:
        type: string
      salePrice:
        type: number
      subProduct:
        items:
          $ref: '#/definitions/models.CreateProductISubProduct'
        type: array
      unit:
        type: string
    required:
    - barcodeId
    - basePrice
    - groupId
    - productName
    - productNameTh
    - salePrice
    - unit
    type: object
  models.CreateProductISubProduct:
    properties:
      baseProductId:
        type: string
      baseProductUse:
        type: integer
    type: object
  models.CreateSaleChannelI:
    properties:
      saleChannel:
        type: string
    required:
    - saleChannel
    type: object
...
  models.DeleteProductI:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  models.DeleteSaleChannelI:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  models.GetStockI:
    properties:
      barcodeId:
        type: string
      productId:
        type: string
      productName:
        type: string
    type: object
  models.GetStockO:
    properties:
      barcodeId:
        type: string
      productId:
        type: string
      productName:
        type: string
      quantity:
        type: integer
    type: object
  models.ProductCreateI:
    properties:
      images:
        items:
          type: string
        type: array
      nameEn:
        type: string
      nameTh:
        type: string
      price:
        $ref: '#/definitions/collectionmodels.Price'
      productId:
        type: string
      productList:
        items:
          $ref: '#/definitions/collectionmodels.ProductList'
        type: array
      shopItem:
        $ref: '#/definitions/collectionmodels.ShopItem'
      shopSku:
        type: string
      skuId:
        type: string
      status:
        type: string
      stock:
        $ref: '#/definitions/collectionmodels.Stock'
      url:
        type: string
    type: object
  models.ProductDeleteI:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  models.ProductUpdateI:
    properties:
      id:
        type: string
      images:
        items:
          type: string
        type: array
      nameEn:
        type: string
      nameTh:
        type: string
      price:
        $ref: '#/definitions/collectionmodels.Price'
      productId:
        type: string
      productList:
        items:
          $ref: '#/definitions/collectionmodels.ProductList'
        type: array
      shopItem:
        $ref: '#/definitions/collectionmodels.ShopItem'
      shopSku:
        type: string
      skuId:
        type: string
      status:
        type: string
      stock:
        $ref: '#/definitions/collectionmodels.Stock'
      url:
        type: string
    type: object
...
  models.ReadProductO:
    properties:
      barcodeId:
        type: string
      basePrice:
        type: number
      groupId:
        type: string
      id:
        type: string
      isBaseProduct:
        type: boolean
      productName:
        type: string
      productNameTh:
        type: string
      productSub:
        items:
          $ref: '#/definitions/models.ReadProductOSub'
        type: array
      salePrice:
        type: number
      unit:
        type: string
    type: object
  models.ReadProductOSub:
    properties:
      baseProductId:
        type: string
      baseProductUse:
        type: integer
      id:
        type: string
      productId:
        type: string
    type: object
  models.ReadSaleChannelI:
    properties:
      saleChannel:
        type: string
    type: object
  models.ReadSaleChannelO:
    properties:
      id:
        type: string
      saleChannel:
        type: string
    type: object
  models.ResponseAPI-any:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  models.ResponseAPI-array_models_GetStockO:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.GetStockO'
        type: array
      message:
        type: string
    type: object
...
  models.ResponseAPI-array_models_ReadProductO:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.ReadProductO'
        type: array
      message:
        type: string
    type: object
  models.ResponseAPI-array_models_ReadSaleChannelO:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/models.ReadSaleChannelO'
        type: array
      message:
        type: string
    type: object
...
  models.ResponseAPI-models_ReadProductO:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/models.ReadProductO'
      message:
        type: string
    type: object
...
  models.ResponseAPI-models_UpdateProductO:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/models.UpdateProductO'
      message:
        type: string
    type: object
...
  models.UpdateProductI:
    properties:
      barcodeId:
        type: string
      basePrice:
        type: number
      groupId:
        type: string
      id:
        type: string
      isBaseProduct:
        type: boolean
      productName:
        type: string
      productNameTh:
        type: string
      salePrice:
        type: number
      subProduct:
        items:
          $ref: '#/definitions/models.UpdateProductISubProduct'
        type: array
      unit:
        type: string
    required:
    - id
    type: object
  models.UpdateProductISubProduct:
    properties:
      baseProductId:
        type: string
      baseProductUse:
        type: integer
      id:
        type: string
    required:
    - id
    type: object
  models.UpdateProductO:
    type: object
  models.UpdateSaleChannelI:
    properties:
      id:
        type: string
      saleChannel:
        type: string
    required:
    - id
    type: object
info:
  contact: {}
  description: This is a docs of Go Backend POS API
  title: Go Backend POS API
  version: "1.0"
paths:
...
  /api/v1/product:
    delete:
      description: to delete product item
      operationId: ProductDelete
      parameters:
      - description: ProductDeleteI
        in: body
        name: ProductDeleteI
        schema:
          $ref: '#/definitions/models.ProductDeleteI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductDelete
      tags:
      - Product
    get:
      description: to read product item by id or product name
      operationId: ProductRead
      parameters:
      - description: productname
        in: query
        name: productname
        type: string
      - description: id
        in: query
        name: id
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-models_ReadProductO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductRead
      tags:
      - Product
    post:
      description: to create product item
      operationId: ProductCreate
      parameters:
      - description: ProductCreateI
        in: body
        name: ProductCreateI
        schema:
          $ref: '#/definitions/models.ProductCreateI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductCreate
      tags:
      - Product
    put:
      description: to update product item
      operationId: ProductUpdate
      parameters:
      - description: ProductUpdateI
        in: body
        name: ProductUpdateI
        schema:
          $ref: '#/definitions/models.ProductUpdateI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-models_UpdateProductO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductUpdate
      tags:
      - Product
...
  /api/v1/productitem:
    delete:
      description: to delete product item
      operationId: ProductItemDelete
      parameters:
      - description: DeleteProductI
        in: body
        name: DeleteProductI
        schema:
          $ref: '#/definitions/models.DeleteProductI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductItemDelete
      tags:
      - ProductItem
    get:
      description: to read product item
      operationId: ProductItemRead
      parameters:
      - description: productName
        in: query
        name: productName
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-array_models_ReadProductO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductItemRead
      tags:
      - ProductItem
    post:
      description: to create product item
      operationId: ProductItemCreate
      parameters:
      - description: CreateProductI
        in: body
        name: CreateProductI
        schema:
          $ref: '#/definitions/models.CreateProductI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductItemCreate
      tags:
      - ProductItem
    put:
      description: to update product item
      operationId: ProductItemUpdate
      parameters:
      - description: UpdateProductI
        in: body
        name: UpdateProductI
        schema:
          $ref: '#/definitions/models.UpdateProductI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductItemUpdate
      tags:
      - ProductItem
  /api/v1/productstock:
    get:
      description: to read product stock
      operationId: ProductStockRead
      parameters:
      - description: GetStockI
        in: body
        name: GetStockI
        schema:
          $ref: '#/definitions/models.GetStockI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-array_models_GetStockO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductStockRead
      tags:
      - ProductStock
    put:
      description: to update product stock
      operationId: ProductStockUpdate
      parameters:
      - description: AddStockI
        in: body
        name: AddStockI
        schema:
          $ref: '#/definitions/models.AddStockI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: ProductStockUpdate
      tags:
      - ProductStock
  /api/v1/salechannel:
    delete:
      description: to delete sale channel
      operationId: SaleChannelDelete
      parameters:
      - description: DeleteSaleChannelI
        in: body
        name: DeleteSaleChannelI
        schema:
          $ref: '#/definitions/models.DeleteSaleChannelI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: SaleChannelDelete
      tags:
      - SaleChannel
    get:
      description: to read sale channel
      operationId: SaleChannelRead
      parameters:
      - description: ReadSaleChannelI
        in: body
        name: ReadSaleChannelI
        schema:
          $ref: '#/definitions/models.ReadSaleChannelI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-array_models_ReadSaleChannelO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: SaleChannelRead
      tags:
      - SaleChannel
    post:
      description: to create sale channel
      operationId: SaleChannelCreate
      parameters:
      - description: CreateSaleChannelI
        in: body
        name: CreateSaleChannelI
        schema:
          $ref: '#/definitions/models.CreateSaleChannelI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: SaleChannelCreate
      tags:
      - SaleChannel
    put:
      description: to update sale channel
      operationId: SaleChannelUpdate
      parameters:
      - description: UpdateSaleChannelI
        in: body
        name: UpdateSaleChannelI
        schema:
          $ref: '#/definitions/models.UpdateSaleChannelI'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseAPI-any'
      summary: SaleChannelUpdate
      tags:
      - SaleChannel
...
securityDefinitions:
  bearerAuth:
    description: 'Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"